+++ GravityCommon/Aseprite.cpp
@@ -42,11 +42,11 @@ bool LoadSpriteSheet(
 
     json_t const* w = json_getProperty(size, "w");
     if (Bad(w, JSON_INTEGER, "meta.size.w")) { return EXIT_FAILURE; }
-    sheet.meta.size.w = json_getInteger(w);
+    sheet.meta.size.w = static_cast<float>(json_getInteger(w));
 
     json_t const* h = json_getProperty(size, "h");
     if (Bad(h, JSON_INTEGER, "meta.size.h")) { return EXIT_FAILURE; }
-    sheet.meta.size.h = json_getInteger(h);
+    sheet.meta.size.h = static_cast<float>(json_getInteger(h));
 
     json_t const* scale = json_getProperty(meta, "scale");
     if (Bad(scale, JSON_TEXT, "meta.scale")) { return EXIT_FAILURE; }
@@ -66,11 +66,11 @@ bool LoadSpriteSheet(
 
         json_t const* from = json_getProperty(tagobj, "from");
         if (Bad(from, JSON_INTEGER, "meta.frameTags.tag.from")) { return EXIT_FAILURE; }
-        newframetag.from = json_getInteger(from);
+        newframetag.from = static_cast<uint32_t>(json_getInteger(from));
 
         json_t const* to = json_getProperty(tagobj, "to");
         if (Bad(to, JSON_INTEGER, "meta.frameTags.tag.to")) { return EXIT_FAILURE; }
-        newframetag.to = json_getInteger(to);
+        newframetag.to = static_cast<uint32_t>(json_getInteger(to));
 
         json_t const* direction = json_getProperty(tagobj, "direction");
         if (Bad(direction, JSON_TEXT, "meta.frameTags.tag.direction")) { return EXIT_FAILURE; }
@@ -93,19 +93,19 @@ bool LoadSpriteSheet(
 
         json_t const* x = json_getProperty(frame, "x");
         if (Bad(x, JSON_INTEGER, "frames.cel.frame.x")) { return EXIT_FAILURE; }
-        newcel.frame.position.x = json_getInteger(x);
+        newcel.frame.position.x = static_cast<float>(json_getInteger(x));
 
         json_t const* y = json_getProperty(frame, "y");
         if (Bad(y, JSON_INTEGER, "frames.cel.frame.y")) { return EXIT_FAILURE; }
-        newcel.frame.position.y = json_getInteger(y);
+        newcel.frame.position.y = static_cast<float>(json_getInteger(y));
 
         json_t const* w = json_getProperty(frame, "w");
         if (Bad(w, JSON_INTEGER, "frames.cel.frame.w")) { return EXIT_FAILURE; }
-        newcel.frame.size.w = json_getInteger(w);
+        newcel.frame.size.w = static_cast<float>(json_getInteger(w));
 
         json_t const* h = json_getProperty(frame, "h");
         if (Bad(h, JSON_INTEGER, "frames.cel.frame.h")) { return EXIT_FAILURE; }
-        newcel.frame.size.h = json_getInteger(h);
+        newcel.frame.size.h = static_cast<float>(json_getInteger(h));
 
         json_t const* rotated = json_getProperty(cel, "rotated");
         if (Bad(rotated, JSON_BOOLEAN, "frames.cel.rotated")) { return EXIT_FAILURE; }
@@ -120,34 +120,34 @@ bool LoadSpriteSheet(
 
         x = json_getProperty(spritesourcesize, "x");
         if (Bad(x, JSON_INTEGER, "frames.cel.spritesourcesize.x")) { return EXIT_FAILURE; }
-        newcel.spritesourcesize.position.x = json_getInteger(x);
+        newcel.spritesourcesize.position.x = static_cast<float>(json_getInteger(x));
 
         y = json_getProperty(spritesourcesize, "y");
         if (Bad(y, JSON_INTEGER, "frames.cel.spritesourcesize.y")) { return EXIT_FAILURE; }
-        newcel.spritesourcesize.position.y = json_getInteger(y);
+        newcel.spritesourcesize.position.y = static_cast<float>(json_getInteger(y));
 
         w = json_getProperty(spritesourcesize, "w");
         if (Bad(w, JSON_INTEGER, "frames.cel.spritesourcesize.w")) { return EXIT_FAILURE; }
-        newcel.spritesourcesize.size.w = json_getInteger(w);
+        newcel.spritesourcesize.size.w = static_cast<float>(json_getInteger(w));
 
         h = json_getProperty(spritesourcesize, "h");
         if (Bad(h, JSON_INTEGER, "frames.cel.spritesourcesize.h")) { return EXIT_FAILURE; }
-        newcel.spritesourcesize.size.h = json_getInteger(h);
+        newcel.spritesourcesize.size.h = static_cast<float>(json_getInteger(h));
 
         json_t const* sourcesize = json_getProperty(cel, "sourceSize");
         if (Bad(sourcesize, JSON_OBJ, "frames.cel.sourceSize")) { return EXIT_FAILURE; }
 
         w = json_getProperty(frame, "w");
         if (Bad(w, JSON_INTEGER, "frames.cel.sourceSize.w")) { return EXIT_FAILURE; }
-        newcel.sourcesize.w = json_getInteger(w);
+        newcel.sourcesize.w = static_cast<float>(json_getInteger(w));
 
         h = json_getProperty(frame, "h");
         if (Bad(h, JSON_INTEGER, "frames.cel.sourceSize.h")) { return EXIT_FAILURE; }
-        newcel.sourcesize.h = json_getInteger(h);
+        newcel.sourcesize.h = static_cast<float>(json_getInteger(h));
 
         json_t const* duration = json_getProperty(cel, "duration");
         if (Bad(duration, JSON_INTEGER, "frames.cel.duration")) { return EXIT_FAILURE; }
-        newcel.duration = json_getInteger(duration);
+        newcel.duration = static_cast<float>(json_getInteger(duration));
 
         sheet.frames.push_back(newcel);
     }
